tools/writing/voice-cloner/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-cloner"
version = "0.1.0"
description = "Analyze writing samples and generate a Claude Code skill capturing the author's voice."
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "PyYAML>=6.0",
    "python-docx>=1.0",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
voice-cloner = "cli:cli"

[tool.setuptools]
py-modules = ["cli", "reader", "analyzer", "synthesizer", "generator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
